\name{daqConfig.NIDAQmx}
\alias{daqConfig.NIDAQmx}
\title{Configure NIDAQmx device.}
\usage{
  daqConfig.NIDAQmx(dev, freq = 1000, nsamples = 100)
}
\arguments{
  \item{dev}{Device.}

  \item{freq}{Sampling reate in Hz.}

  \item{nsamples}{Number of samples to read.}
}
\description{
  Configure NIDAQmx device.
}

